Back button from SettingsActivity applies theme
After the theme has been changed from the settings menu and the software
back button is pressed, the theme wouldn't update properly.

The theme is applied while an activity is created. However, the software back
button doesn't recreate the previous activity, it just resumes it. This commit
makes sure that the activity is re-created only if required. If there is no
theme change, the previous activity will be resumed.

Sequence of creation of the NotificationService has been modified to allow for
early exit if need be.

Close: #111