Added the ability to share modules without having to download them
One can now share a link using any text message app and that link can be opened by the receiving user
using the app which will open the file in the browser using the receiving user's
token. If they are not enrolled in the course of which the shared module is a part,
they are shown a Toast which says that they need to enroll in this course in order to
view that file.

close #137