Create FolderModuleFragment to support Folder modules
Folder modules are a frequent occurence on CMS, which the app did not
have support for.

Folder modules are essentially a folder of files. Earlier, the app would
download all files inside a folder and would open up the first file of
that folder.

The fragment allows for the listing of all files inside the folder
module.

Close #23