AlertDialog to view File properties
Module content and discussion attachements come with metadata like the
file size and last modified epoch. Such information maybe relavant to a
user.

An AlertDialog is used to show the file size and the last modified
properties of a file. Both are shown in a easy-to-read way without
hindering UX.

Close #142