Round More Options BottomSheetDialogFragment corners
Apart from what the title says, this commit also changes the parent of
the `AppTheme` style. BottomSheetDialogFragment being a material class,
a couple of attributes needed to be set in our styles. These are set by
the new parent Material Components theme, removing the need for us to move
these our to our style. The lack of these would cause an InflateException
RTE.

This change also meant that the `onCreateView` inflater context jugaad
that was done in `MoreOptionsFragment` is no longer required. I still
have no clue why that was a problem to be begin with.

Overriding the default BottomSheetDialog themes also means that we don't
have to worry about animations in our code or styles.

This small change has resulted in a massive simplification of
our codebase and styles.