Cleanup: Disable app backups
The app doesn't use SharedPreferences for anything important. Neither
does it use app storage for files. Disable app backups removed the need
to provide a `fullBackupDescriptor` for the app.

Even if it were allowed, other than preferences for notifications and
app theme, all other SharedPreference uses is for login and
authentication and that should not be backedup anyways.

Note: Ultimately, this descision was taken to get rid of an annoying
lint error which required us to provide a backup descriptor which
excluded all categories of app data. It wasn't worthwhile making such a
change, espescially when setting up the from a fresh install isn't all
that hard anyways.