Cleanup: CourseSectionFragment and associated layouts
The CourseSectionFragment is a complete shitshow. This patch cleans it
up and optimizes the performance of the fragment.

1. All course sections are added to a nested scroll view. This
kills performance if there are a lot of course sections. A systrace
revealed that it took 5 seconds for the fragment to loadup for a course
with a lot of fragments.

2. Each course section had a RecylerView list for the modules
under that section. This didn't take advantage of the RecylerView's
caching and recycling features. A normal ListView would have sufficed.

My initial attempt at a cleanup involved using nested RecyclerViews.
However, this wasn't the way to go since RecylerViews don't support
being nested in each other.

I concluded the best way to go about a cleanup would be to use a single
RecyclerView for both course sections and course modules, collectively
called Course Contents. A unified list of both is presented using the
RecyclerView with the associated adapter taking care of inflating the
correct layouts. The external dependency AdapterDelegates was added to
make it more easier to support multiple view types as well as add new
ones in the future.

Further more, course sections are filtered out early on instead of doing
so while binding. This removes the need for the LayoutManager to do a
measure pass of such sections just to realize they are all gone.

These changes has resulted in a bug with the collapsible descriptions
of modules. For some weird reason, the layout updates don't result in a
callback when scrolling through the list. However, I considered it
unnecessary to fix this bug in this commit for the simple reason that
the current implementation is extremley difficult to wrap our heads
around. I intend to clean this up as well by introducing a new text
widgest that automatically deals with collpasing and display long text
content.