Add: CollapsibleTextView for collapsing and expanding text
The current collapsing logic is extremley convoluted. It sets text on a
standard TextView and waits for `onGlobalLayout()` to count the number
of lines and then shrinks or collapses the text view. Further more, the
way it comes up with the number of characters to shown when collapsed is
a total nightmare.

This patch introduces a custom TextView that removes the need for a
measure and layout pass before counting the number of lines taken by the
text. Instead, this is done in the measure pass itself, requiring one
additional measure pass if anything changes.

This should net us a small performance gain. More importantly, this
patch fixes any bugs that were introduced by switching to a RecylerView
adapter that merged course sections and coures contents.