Cleanup: crux.bphc.cms.helper package
This package was basically the dumping ground for any classes not fit in
the other (original) packages. Sometimes, classes meant to be in another
package was inside helper (eg: UserDetail)

This PR aims to resolve much of the structuring problem as well as lint
issues. This is definitley not the end of the cleanup for this package
and its contents, existing as well to-be relocated contents