Replace TravisCI with GitHub Actions
GitHub Actions is integrated into GitHub and will makes lives easier
than if with TravisCI.

The default TravisCI job was to build the app. The Actions workflow is
to assemble APKs of all build variants. The debug APK is available for
download as an artifact.

An additional lint job has also been added which will upload the lint
reports as an artifact once the job is done.