Use coroutines when refreshing courses in MyCoursesFragment
The API responses were being handled on the main thread, which caused
jank and maybe even ANRs when a large number of responses were to be
handled. Coroutines running on threadpools help too fix this problem.

However, due to the fact that these were running on thread pools,
a new realm instance had to be tied to each coroutine instance. Further
more, API requests also had to be made synchronoulsy since async
requests are posted on the main thread by Retrofit.

Close #225