Switch to WorkManager from JobScheduler for NotificationService
WorkManager is part of Jetpack and is the Android recommended way to
schedule periodic work in the background.

This commit cleans up the erstwhile NotificationService.java file,
Kotlin-ifying and refactoring it. InboxStyle has been replaced with
BigTextStyle notifications on Oreo and above.

In cases where a logout happens due to InvalidToken or any other reason,
a notification is shown to the user. This closes #241.