Refactor Course Enrol code flow
Course Enrolment involves searching for a course from all available
courses followed by enrolment. The icon in the bottom naviation was that
of search icon with text of 'Search Course'. The icon did not imply the
ability to enrol to courses.

This patch fixes the aforementioned navigational issue and renames
`SearchCourseFragment` to `SearchCourseFragmentForEnrolFragment`, which
is more appropriate.