Handle module and section push notifs when app in background
When app is backgrounded, the system simply creates a notification. Once
the notification is clicked, the app's default launcher is called with
the notification data passed as intent extras.

This patch changes the default launcher to MainActivity and has it
launch CourseDetailActivity and handle the notification data. This will
reduce the loading time experienced by the user once the notification is
clicked. Otherwise, the app will jump from TokenActivity -> MainActivity
-> CourseDetailActivity.

Urls embeded in the push notif is used to determine the context of the
notification and open up the appropriate screen and scroll to the new
module/discussion.