Add: Option to unenrol from courses
This commit introduces a bunch of code to add the ability to unenrol
students from courses. The method is hacky, involving creating an HTTP
session using a private token, followed by emulating a HTML form submit
to do the unenrolment. Unfortunately, Moodle doesn't have a webservice
to allow for unenrolment. :(